fix structure to a default nextjs
